Better coverage with mypy fixed and small refactors
